+++ src/components/Accordion/Accordion.tsx
@@ -23,6 +23,9 @@ export interface AccordionProps {
   id?: string;
   /** Whether the accordion is open by default only on its initial rendering */
   isDefaultOpen?: boolean;
+  /** Determines the max height of the accordion panel; if content is larger
+   * than panelMaxHeight, a scroll bar appears */
+  panelMaxHeight?: string;
 }
 
 /**
@@ -41,7 +44,11 @@ const getIcon = (isExpanded = false, index: number, id: string) => {
  * array. This automatically creates the `AccordionButton` and `AccordionPanel`
  * combination that is required for the Chakra `Accordion` component.
  */
-const getElementsFromData = (data: AccordionDataProps[] = [], id: string) => {
+const getElementsFromData = (
+  data: AccordionDataProps[] = [],
+  id: string,
+  panelMaxHeight: string
+) => {
   const colorMap = {
     default: "ui.white",
     warning: "ui.status.primary",
@@ -60,9 +67,16 @@ const getElementsFromData = (data: AccordionDataProps[] = [], id: string) => {
           id={`${id}-panel-${index}`}
           key={index}
           dangerouslySetInnerHTML={{ __html: content.panel }}
+          maxHeight={panelMaxHeight}
+          overflow="auto"
         />
       ) : (
-        <AccordionPanel id={`${id}-panel-${index}`} key={index}>
+        <AccordionPanel
+          id={`${id}-panel-${index}`}
+          key={index}
+          maxHeight={panelMaxHeight}
+          overflow="auto"
+        >
           {content.panel}
         </AccordionPanel>
       );
@@ -124,7 +138,13 @@ const getElementsFromData = (data: AccordionDataProps[] = [], id: string) => {
  */
 export const Accordion = chakra(
   forwardRef<HTMLDivElement, AccordionProps>((props, ref?) => {
-    const { accordionData, id, isDefaultOpen = false, ...rest } = props;
+    const {
+      accordionData,
+      id,
+      isDefaultOpen = false,
+      panelMaxHeight,
+      ...rest
+    } = props;
 
     // Pass `0` to open the first accordion in the 0-index based array.
     const openFirstAccordion = isDefaultOpen ? [0] : undefined;
@@ -137,7 +157,7 @@ export const Accordion = chakra(
         ref={ref}
         {...rest}
       >
-        {getElementsFromData(accordionData, id)}
+        {getElementsFromData(accordionData, id, panelMaxHeight)}
       </ChakraAccordion>
     );
   })